changed html file names to be more intuitive
All the html files were named index.html under their respective folders
and thought that was not a good way to name all our html files. Contact
was already contact.html and so I went ahead and changed it for the
others. However, I've been having an issue with blog because it seems to
be connected somehow to pagination. Will look into.