Edited blog page to show posts as excerpts
The blog page had several issues. It was showing the entire post rather
than giving an excerpt of the posts. It was also paginated, which is why
the nav bar link to the blog page was hardcoded. This was changed so I
removed the hardcoded link in the nav menu. Made each post to be
displayed as its own page using the post.html template.