Refactored Blog to have pagination capacity
In order to provide pagination capability, I changed the conifg file and
updated the data structure. Jekyll requires paginated pages to be titled
index.html and therefore I moved the blog.html file that requires
pagination in to it's own blog folder and renamed to index.html. Added
links to older and newer articles as navigation for paginated pages.
This also required to nav links to be hardcoded, so made the necessary
changes.