Updated blog design and added pages for categories
The blog has been redesigned to have a side bar on the right along with
a different layout for listing posts. I also made separate pages to sort
blogs by categories. The sidebar has also been added to the individual
posts. Categories in posts have been edited to be consistent with the
lowercase spelling of all categories.