Tinified profiles images and added thumbnails
To make our profile pages load faster, I tinified the profile images of
our tutors and then also created thumbnails for them. While doing so, I
added Carlin's image and changed kimball's image from png to jpg to be
consistent with the rest. Currently, the thumbnails are at minimum of
400x400.